backend/requirements.txt
alembic==1.18.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
attrs==25.4.0
bcrypt==5.0.0
billiard==4.2.4
build==1.4.0
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.5
chromadb==1.5.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
durationpy==0.10
fastapi==0.128.0
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.2.0
googleapis-common-protos==1.73.0
greenlet==3.3.1
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
kubernetes==35.0.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
numpy==2.4.3
oauthlib==3.3.1
onnxruntime==1.24.3
openai==2.28.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
overrides==7.7.0
packaging==26.0
prompt_toolkit==3.0.52
protobuf==6.33.5
psycopg2-binary==2.9.11
pybase64==1.4.3
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pypdf==5.3.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.5.0
websocket-client==1.9.0
websockets==16.0
wheel==0.46.3
zipp==3.23.0
